Refactor imports in components to use absolute paths for better module organization and consistency.
